% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDFLColsMax.R
\name{get.DFL.cols.max}
\alias{get.DFL.cols.max}
\title{Retrieve max value of each col in a data frame list}
\usage{
get.DFL.cols.max(data_frame_list, cols_to_compute)
}
\arguments{
\item{data_frame_list}{List containing the dataframes to analyse.}

\item{cols_to_compute}{Vector with the columns to analyse.}
}
\value{
Returns a dataframe with the columns name and their maximum for each
dataframe in the specified list.
}
\description{
Retrieve max value of each col in a data frame list
}
