% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDFLColsMax.R
\name{getDFLColsMax}
\alias{getDFLColsMax}
\title{Retrieve max value for specified columns in a data frame list.}
\usage{
getDFLColsMax(dataFrameList, colsToCompute)
}
\arguments{
\item{dataFrameList}{List containing the dataframes to analyse.}

\item{colsToCompute}{Vector with the columns to analyse.}
}
\value{
Returns a dataframe with the columns name and their maximum for each
dataframe in the specified list.
}
\description{
Retrieve max value for specified columns in a data frame list.
}
\examples{
getDFLColsMax(myList, c("Time", "Distance"))
getDFLColsMax(myList, columnsList)

}
