% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDFLColsMean.R
\name{getDFLColsMean}
\alias{getDFLColsMean}
\title{Compute the mean value of specified columns for a list of dataframes.}
\usage{
getDFLColsMean(dataFrameList, colsToCompute)
}
\arguments{
\item{dataFrameList}{A list containing the data frames to analyse.}

\item{colsToCompute}{A vector with the columns to analyse.}
}
\value{
Returns a dataframe with the columns name and their means for each
dataframe in the specified list.
}
\description{
Compute the mean value of specified columns for a list of dataframes.
}
\examples{
getDFLColsMean(myList, c("Time", "Distance"))
getDFLColsMean(myList, columnsList)

}
