% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDFLColsMedian.R
\name{getDFLColsMedian}
\alias{getDFLColsMedian}
\title{Compute the median value of specified columns for a list of dataframes.}
\usage{
getDFLColsMedian(dataFrameList, colsToCompute)
}
\arguments{
\item{dataFrameList}{List containing the dataframes to analyse.}

\item{colsToCompute}{A vector with the columns to analyse.}
}
\value{
Returns a dataframe with the columns name and their means for each
dataframe in the specified list.
}
\description{
Compute the median value of specified columns for a list of dataframes.
}
\examples{
getDFLColsMedian(myList, c("Time", "Distance"))
getDFLColsMedian(myList, columnsList)

}
