% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDFLColsSum.R
\name{getDFLColsSum}
\alias{getDFLColsSum}
\title{Compute the sums of specified columns for a list of dataframes.}
\usage{
getDFLColsSum(dataFrameList, colsToCompute)
}
\arguments{
\item{dataFrameList}{List containing the dataframes to analyse.}

\item{colsToCompute}{A vector with the columns to analyse.}
}
\value{
Returns a dataframe with the columns name and their sums for each
dataframe in the specified list.
}
\description{
Compute the sums of specified columns for a list of dataframes.
}
\examples{
getDFLColsSum(myList, c("Time", "Distance"))
getDFLColsSum(myList, columnsList)

}
