% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDFLColsVals.R
\name{getDFLColsVal}
\alias{getDFLColsVal}
\title{Retrieve values in each data frame of a list}
\usage{
getDFLColsVal(dataFrameList, colsToGet, dataFrameSelection = "all")
}
\arguments{
\item{dataFrameList}{List containing the dataframes to analyse.}

\item{colsToGet}{A vector with the columns to retrieve.}

\item{dataFrameSelection}{A list of dataframes to analyse in dataFrameList.
By default all dataframes are used.}
}
\value{
Returns a dataframe with the columns name (column and dataframe name)
and their values for each dataframe in the specified list.
}
\description{
Retrieve values in each data frame of a list
}
\examples{
getDFLColsVal(myList, c("VE", "VO2"))
getDFLColsVal(myList, c("VE", "VO2"), c("DF1", "DF2"))

}
