% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDFLRowsMeanRelative.R
\name{getDFLRowsMeanRel}
\alias{getDFLRowsMeanRel}
\title{Title}
\usage{
getDFLRowsMeanRel(dataFrameList, colsToCompute)
}
\arguments{
\item{colsToCompute}{}
}
\value{

}
\description{
Title
}
\examples{

}
